+++ app/res/web2board/linux/res/platformioWorkSpaceSkeleton/lib/BitbloqMBot/BitbloqMBot.cpp
@@ -1,1 +1,0 @@
-/*Deprecated see BitbloqMBotV2*/